requirements.txt
# Requirements for the Python project
gradio>=4.0.0
openai>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0